Move ByteStoreProvider trait to remote_provider_traits
