+++ src/rust/engine/fs/store/Cargo.toml
@@ -28,6 +28,7 @@ madvise = { workspace = true }
 parking_lot = { workspace = true }
 prost = { workspace = true }
 prost-types = { workspace = true }
+remote_provider_traits = { path = "../../remote_provider/remote_provider_traits" }
 serde = { workspace = true }
 serde_derive = { workspace = true }
 sharded_lmdb = { path = "../../sharded_lmdb" }
