+++ src/rust/engine/fs/store/src/remote.rs
@@ -6,15 +6,14 @@ use std::sync::Arc;
 use std::time::{Duration, Instant};
 
 use async_oncecell::OnceCell;
-use async_trait::async_trait;
 use bytes::Bytes;
 use futures::Future;
 use hashing::Digest;
 use log::Level;
 use protos::gen::build::bazel::remote::execution::v2 as remexec;
 use remexec::ServerCapabilities;
+use remote_provider_traits::{ByteStoreProvider, LoadDestination};
 use tokio::fs::File;
-use tokio::io::{AsyncSeekExt, AsyncWrite};
 use workunit_store::{in_workunit, ObservationMetric};
 
 mod reapi;
@@ -24,30 +23,6 @@ mod reapi_tests;
 pub mod base_opendal;
 #[cfg(test)]
 mod base_opendal_tests;
-
-#[async_trait]
-pub trait ByteStoreProvider: Sync + Send + 'static {
-  /// Store the bytes readable from `file` into the remote store
-  async fn store_file(&self, digest: Digest, file: File) -> Result<(), String>;
-
-  /// Store the bytes in `bytes` into the remote store, as an optimisation of `store_file` when the
-  /// bytes are already in memory
-  async fn store_bytes(&self, digest: Digest, bytes: Bytes) -> Result<(), String>;
-
-  /// Load the data stored (if any) in the remote store for `digest` into `destination`. Returns
-  /// true when found, false when not.
-  async fn load(
-    &self,
-    digest: Digest,
-    destination: &mut dyn LoadDestination,
-  ) -> Result<bool, String>;
-
-  /// Return any digests from `digests` that are not (currently) available in the remote store.
-  async fn list_missing_digests(
-    &self,
-    digests: &mut (dyn Iterator<Item = Digest> + Send),
-  ) -> Result<HashSet<Digest>, String>;
-}
 
 // TODO: Consider providing `impl Default`, similar to `super::LocalOptions`.
 #[derive(Clone)]
@@ -107,29 +82,6 @@ pub struct ByteStore {
 impl fmt::Debug for ByteStore {
   fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
     write!(f, "ByteStore(name={:?})", self.instance_name)
-  }
-}
-
-/// Places that write the result of a remote `load`
-#[async_trait]
-pub trait LoadDestination: AsyncWrite + Send + Sync + Unpin + 'static {
-  /// Clear out the writer and start again, if there's been previous contents written
-  async fn reset(&mut self) -> std::io::Result<()>;
-}
-
-#[async_trait]
-impl LoadDestination for tokio::fs::File {
-  async fn reset(&mut self) -> std::io::Result<()> {
-    self.rewind().await?;
-    self.set_len(0).await
-  }
-}
-
-#[async_trait]
-impl LoadDestination for Vec<u8> {
-  async fn reset(&mut self) -> std::io::Result<()> {
-    self.clear();
-    Ok(())
   }
 }
