+++ src/rust/engine/fs/store/src/remote/base_opendal_tests.rs
@@ -6,11 +6,12 @@ use std::time::Duration;
 use bytes::Bytes;
 use grpc_util::tls;
 use opendal::services::Memory;
+use remote_provider_traits::ByteStoreProvider;
 use testutil::data::TestData;
 use testutil::file::mk_tempfile;
 
 use super::base_opendal::Provider;
-use super::{ByteStoreProvider, RemoteOptions};
+use super::RemoteOptions;
 
 const BASE: &str = "opendal-testing-base";
