+++ src/rust/engine/remote_provider/remote_provider_traits/Cargo.toml
@@ -6,3 +6,7 @@ authors = ["Pants Build <pantsbuild@gmail.com>"]
 publish = false
 
 [dependencies]
+async-trait = { workspace = true }
+bytes = { workspace = true }
+hashing = { path = "../../hashing" }
+tokio = { workspace = true, features = ["fs"] }
